# Faster command module loading & tab completion performance