--- a/src/azure-cli-core/azure/cli/core/parser.py
+++ b/src/azure-cli-core/azure/cli/core/parser.py
@@ -87,7 +87,6 @@ class AzCliCommandParser(argparse.ArgumentParser):
                                         command=command_name,
                                         _validators=argument_validators,
                                         _parser=command_parser)
-        enable_autocomplete(self)
 
     def _get_subparser(self, path):
         """For each part of the path, walk down the tree of