--- a/src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/__init__.py
+++ b/src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/__init__.py
@@ -3,10 +3,10 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 #---------------------------------------------------------------------------------------------
 
-#pylint: disable=unused-import
+import azure.cli.command_modules.acr._help #pylint: disable=unused-import
 
-import azure.cli.command_modules.acr._help
-import azure.cli.command_modules.acr._params
-import azure.cli.command_modules.acr.custom
-import azure.cli.command_modules.acr.credential
-import azure.cli.command_modules.acr.repository
+def load_params(_):
+    import azure.cli.command_modules.acr._params #pylint: disable=redefined-outer-name
+
+def load_commands():
+    import azure.cli.command_modules.acr.commands #pylint: disable=redefined-outer-name