… doc/authoring_command_modules/README.md
… doc/authoring_command_modules/authoring_commands.md
… src/azure-cli-core/azure/cli/core/tests/test_command_registration.py
… src/azure-cli-core/azure/cli/core/tests/test_generic_update.py
… src/azure-cli-core/azure/cli/core/tests/test_help.py
… src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/custom.py
… src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/repository.py
… src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/__init__.py
… src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/commands.py
… src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/__init__.py
… src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_client_factory.py
… src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_params.py
… src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/commands.py
… src/command_modules/azure-cli-cloud/azure/cli/command_modules/cloud/__init__.py
… src/command_modules/azure-cli-cloud/azure/cli/command_modules/cloud/commands.py
… src/command_modules/azure-cli-component/azure/cli/command_modules/component/__init__.py
… src/command_modules/azure-cli-component/azure/cli/command_modules/component/commands.py
… src/command_modules/azure-cli-configure/azure/cli/command_modules/configure/__init__.py
… src/command_modules/azure-cli-configure/azure/cli/command_modules/configure/commands.py
… src/command_modules/azure-cli-configure/azure/cli/command_modules/configure/custom.py
… src/command_modules/azure-cli-container/azure/cli/command_modules/container/__init__.py
… src/command_modules/azure-cli-container/azure/cli/command_modules/container/commands.py
… src/command_modules/azure-cli-context/azure/cli/command_modules/context/__init__.py
… src/command_modules/azure-cli-context/azure/cli/command_modules/context/commands.py
… src/command_modules/azure-cli-feedback/azure/cli/command_modules/feedback/__init__.py
… src/command_modules/azure-cli-feedback/azure/cli/command_modules/feedback/commands.py
… src/command_modules/azure-cli-feedback/azure/cli/command_modules/feedback/custom.py
… src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/__init__.py
… src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/commands.py
… src/command_modules/azure-cli-keyvault/azure/cli/command_modules/keyvault/__init__.py