Closes #835

This commit implements this for the following modules only:
- component
- vm

(other modules to come in later commits)

Although the PR is DO NOT MERGE, it is open for comments before I extend this strategy out to other command modules.
After completing the change for other command modules, I'll remove the DO NOT MERGE.

<!-- Reviewable:start -->
---

This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/azure/azure-cli/1059)

<!-- Reviewable:end -->