@tjprescott 2016-10-11
Also, what are the speed gains with this approach?