@derekbekoe 2016-10-11
Perf improvements once all modules have been converted:
- Tab completion is faster (0.5 secs -> 0.2 secs)
- The CLI loads faster overall (1.5 secs -> 0.4 secs)

The best thing about this change is that as we add more command modules, overall performance won't take as much of a hit as we do right now.

Main reason is because we defer the importing/loading of command handlers until when the command is actually going to be invoked. Also, params are loaded until needed and our params use the SDK heavily for choice lists, etc.