@derekbekoe 2016-10-11
I'll update the command authoring docs also once this is finalized.