@tjprescott 2016-10-11
The command authoring docs changes would help with the review as it is hard to fully understand what onboarding commands used to look like compared to what it would look like now. The new way seems to be a lot more obtuse in the way you add commands.