@derekbekoe 2016-10-13
@tjprescott, @yugangw-msft, @johanste 
I've made the suggested change to remove _operation_factory.py files.

Now, we use strings to define the handler for a command.
For example:
```
cli_command(__name__, 'vm show', 'azure.mgmt.compute.operations.virtual_machines_operations#VirtualMachinesOperations.get', cf_vm)
```

Please take a look.